LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/pixel-manipulation/2d.imageData.get.unaffected-expected.txt

2d.imageData.get.unaffected

getImageData() is not affected by context state


PASS getImageData() is not affected by context state 
